+++ src/app/styles/input-style/input-style.component.ts
@@ -17,10 +17,21 @@ export class InputStyleComponent extends BasicStyleComponent implements OnInit {
         super();
     }
 
-    ngOnInit() { }
+    ngOnInit() { 
+        console.log(this.style);
+    }
 
     formatDate(controlName){
         this.parentForm.controls[controlName].setValue(this.parentForm.controls[controlName].value.split('T')[0]);
     }
+
+    formatDateTime(controlName){
+        if(this.parentForm.controls[controlName].value.includes('+')){
+            let time = this.parentForm.controls[controlName].value.split('+')[0];
+            console.log(time.split('T')[0] + " " + time.split('T')[1]);
+            this.parentForm.controls[controlName].setValue(time.split('T')[0] + " " + time.split('T')[1]);
+        }
+    }
+
 
 }